Changed the navbar so you can tell what page you're on. Black background,yellow text, not sure how else to do it unless we want to make another css sheet.
Groups.html, properly made a form to display groups and join said group, it passes the group_id and the user_name is in the session, so when ready we can take that username and add it to list of people in the group.

Views, just added a post method that deals with 'adding' a user to a group. Basically just decrements the 'spots available' when the button is pressed making it seem like someone 'joined'

models, added "spots available" to group vew