updated navbar with new 'joined groups.html' file
This joined groups file will display what groups the current user has joined. that work has been done accordingly in the views even if user hasn't joined

Groups view now will legitimately display a list of users when you join a group, and if the group is full it will tell you that it's full, if you already joined then it will tell you that you joined and just display who's in the group and tell you to go back and join a new one.

UPdated models to add a manytomany field so that many users can join a group, right now if the 'spotsavailable' gets to 0 you can't join, we can ignore this or just change/update it in the admin side.

and updated the navbar in proper files.

Shouldn't be a db.sqlite3 file in here so yall should be safe